utils_api/requirements.txt
fastapi
uvicorn[standard]
jsonschema
json-schema-to-pydantic
httpx